Move model definitions into separate file
